+++ src/clj_commons/pretty/spec.clj
@@ -8,10 +8,8 @@
 (s/def ::positive-integer (s/and integer? pos?))
 
 
-(s/def ::single-character (s/or
-                            :char char?
-                            :string (s/and string?
-                                           #(= 1 (count %)))))
+(s/def ::single-character (s/and string?
+                                 #(= 1 (count %))))
 
 ;; clj-commons.ansi:
 
@@ -77,7 +75,11 @@
                                     ::ann/nib]))
 
 (s/def ::ann/spacing #{:tall :compact :minimal})
-(s/def ::ann/marker ::single-character)
+(s/def ::ann/marker (s/or
+                      :basic ::single-character
+                      :fn fn?
+                      :tri (s/and string? #(= 3 (count %)))))
+
 (s/def ::ann/bar ::single-character)
 (s/def ::ann/nib ::ansi/composed-string)
 
@@ -86,7 +88,8 @@
 (s/def ::ann/annotation (s/keys :req-un [::ann/message
                                          ::ann/offset]
                                 :opt-un [::ann/length
-                                         ::ansi/font]))
+                                         ::ansi/font
+                                         ::ann/marker]))
 
 (s/def ::ann/message ::ansi/composed-string)
 (s/def ::ann/offset ::nonneg-integer)