+++ CHANGES.md
@@ -3,6 +3,7 @@
 - Default style for `print-table` can be overridden via a dynamic var
 - Use a vertical bar (│) not a pipe character (|) as column separator in binary output
 - New `clj-commons.pretty.nrepl` namespace to set up pretty inside nREPL
+- Function `clj-commons.pretty.repl/main` meant for use with `clj -X` to wrap another function
 
 ## 3.4.1 -- 23 Jun 2025